+++ frb_example/integrate_third_party/android/app/src/main/kotlin/com/example/frb_example_integrate_third_party/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.integrate_third_party
+package com.example.frb_example_integrate_third_party
 
 import io.flutter.embedding.android.FlutterActivity
