+++ frb_example/integrate_third_party/macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = integrate_third_party
+PRODUCT_NAME = frb_example_integrate_third_party
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.integrateThirdParty
+PRODUCT_BUNDLE_IDENTIFIER = com.example.frbExampleIntegrateThirdParty
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2024 com.example. All rights reserved.
