+++ frb_example/integrate_third_party/pubspec.yaml
@@ -1,4 +1,4 @@
-name: integrate_third_party
+name: frb_example_integrate_third_party
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -35,7 +35,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  rust_lib_integrate_third_party:
+  rust_lib_frb_example_integrate_third_party:
     path: rust_builder
   flutter_rust_bridge:
     path: ../../frb_dart
