+++ frb_example/integrate_third_party/rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust_lib_integrate_third_party"
+name = "rust_lib_frb_example_integrate_third_party"
 version = "0.1.0"
 edition = "2021"
