+++ frb_example/integrate_third_party/rust_builder/android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'rust_lib_integrate_third_party'
+rootProject.name = 'rust_lib_frb_example_integrate_third_party'
