+++ frb_example/integrate_third_party/rust_builder/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.flutter_rust_bridge.rust_lib_integrate_third_party">
+  package="com.flutter_rust_bridge.rust_lib_frb_example_integrate_third_party">
 </manifest>
