+++ frb_example/integrate_third_party/web/index.html
@@ -23,13 +23,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="integrate_third_party">
+  <meta name="apple-mobile-web-app-title" content="frb_example_integrate_third_party">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>integrate_third_party</title>
+  <title>frb_example_integrate_third_party</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
