+++ frb_example/integrate_third_party/windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "integrate_third_party" "\0"
+            VALUE "FileDescription", "frb_example_integrate_third_party" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "integrate_third_party" "\0"
+            VALUE "InternalName", "frb_example_integrate_third_party" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2024 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "integrate_third_party.exe" "\0"
-            VALUE "ProductName", "integrate_third_party" "\0"
+            VALUE "OriginalFilename", "frb_example_integrate_third_party.exe" "\0"
+            VALUE "ProductName", "frb_example_integrate_third_party" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
